refactor DropZone + paste
